pyproject.toml
[tool.poetry]
name = "revtool"
version = "0.1.0"
description = ""
authors = ["harrier <harrier@b6a.black>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pwntools = "^4.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
